Add Item, Manage Items

+++ MP GUI VERSION/MPGuiVersion/MPGuiVersion/Base Classes/Item.cs
@@ -9,15 +9,15 @@ namespace MPGuiVersion
         private int itemID;
         private string itemName;
         private int quantity;
-        private float unitPrice;
-        private float totalPrice;
+        private double unitPrice;
+        private double totalPrice;
         private string description;
 
         public int ItemID { get => itemID; set => itemID = value; }
         public string ItemName { get => itemName; set => itemName = value; }
         public int Quantity { get => quantity; set => quantity = value; }
-        public float UnitPrice { get => unitPrice; set => unitPrice = value; }
-        public float TotalPrice { get => (unitPrice * quantity); set => totalPrice = value; }
+        public double UnitPrice { get => unitPrice; set => unitPrice = value; }
+        public double TotalPrice { get => (unitPrice * quantity); set => totalPrice = value; }
         public string Description { get => description; set => description = value; }
     }
 }
